refacto(archi): move Auth
