+++ config/packages/security.yaml
@@ -18,7 +18,7 @@ security:
         main:
             lazy: true
             provider: app_user_provider
-            custom_authenticator: App\Domain\Auth\Authenticator\LoginFormAuthenticator
+            custom_authenticator: App\Auth\Domain\Authenticator\LoginFormAuthenticator
             logout:
                 path: !php/const App\Domain\Routing\ValueObject\RouteName::LOGOUT
                 # where to redirect after logout
