+++ src/Cli/Command/AdminCreate.php
@@ -2,7 +2,7 @@
 
 namespace App\Cli\Command;
 
-use App\Domain\Auth\User;
+use App\Auth\Domain\User;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
